third_party/WebKit/LayoutTests/fast/files/blob-close-revoke-expected.txt

Test the Blob.close() method, revoking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test that dereferencing URLs referring to closed Blobs fail.
PASS Error triggered on loading image from closed Blob.
Test that dereferencing URLs referring to closed Blobs fail, workers.
PASS Worker loaded.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
PASS Error triggered on starting Worker from a closed Blob.
Test creating object URLs on closed Blobs
PASS window.URL.createObjectURL(blob); threw exception InvalidStateError: Failed to execute 'createObjectURL' on 'URL': Blob has been closed..
PASS successfullyParsed is true

TEST COMPLETE
